requirements.txt
boto3==1.26.90
grpcio
wand
numpy
tensorflow==2.12.0
opencv-python
Pillow
protobuf==3.20.0
imageai
torch
torchvision